backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
redis
cryptography
google-generativeai
anthropic
openai
pymupdf
pillow
python-dotenv
